refactor, add a basic test to test_practice and the curious addition of __init__ file??
